+++ NEWS.md
@@ -2,6 +2,9 @@
 
 brio is new in Imports, replacing the use of the Suggested package readr, in `bq_table_download()` (@AdeelK93, #462).
 
+* `bq_table_expiration()` - allows to patch table (@meztez, #467) with expiration.
+
+
 # bigrquery 1.4.0
 
 * `bq_table_download()` has been heavily refactored (#412):